+++ CHANGELOG.md
@@ -10,6 +10,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
  - Added Change log document
 ### Changed
  - Added description on how to use the containerized version of Gens
+ - Replaced print statements with logging to stderr
 ### Fixed
  - Replaced depricated `update()` with `update_one()` in `update_transcripts.py`.